Extract components from src/Tickets/index.js
